completed checkout for order logic removed completely active order and
completed order repos merged into a signle order repo

+++ src/main/java/com/ticketing/infrastructure/gateway/StubPaymentGateway.java
@@ -1,10 +1,12 @@
-package com.ticketing.infrastructure.gateway;
+package  com.ticketing.infrastructure.gateway;
 
-import com.ticketing.domain.gateway.IPaymentGateway;
+import com.ticketing.infrastructure.Interface.*;
+
 import com.ticketing.domain.gateway.PaymentDetails;
 import com.ticketing.domain.gateway.PaymentResult;
 import com.ticketing.domain.gateway.RefundResult;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public class StubPaymentGateway implements IPaymentGateway {
@@ -16,7 +18,7 @@ public class StubPaymentGateway implements IPaymentGateway {
     }
 
     @Override
-    public PaymentResult charge(double amount, PaymentDetails details) {
+    public PaymentResult charge(BigDecimal amount, PaymentDetails details) {
         if (shouldFail) {
             return PaymentResult.failed("Payment declined by issuer.");
         }
@@ -31,3 +33,5 @@ public class StubPaymentGateway implements IPaymentGateway {
         return RefundResult.successful("REF-" + UUID.randomUUID().toString().substring(0, 8));
     }
 }
+
+
